// src/app/[vehicle]/types.d.ts
export interface Option {
  codigo: string;
  nome: string;
}

export interface VehiclePrice {
  TipoVeiculo: number;
  Valor: string;
  Marca: string;
  Modelo: string;
  AnoModelo: number;
  Combustivel: string;
  CodigoFipe: string;
  MesReferencia: string;
  SiglaCombustivel: string;
}